Refactor to avoid mixin classes
