+++ coloraide/__init__.py
@@ -1,7 +1,6 @@
 """ColorAide Library."""
 from .__meta__ import __version_info__, __version__  # noqa: F401
-from .color import Color
-from .color.match import ColorMatch
+from .color import Color, ColorMatch
 from .color.interpolate import Piecewise, Lerp
 from .util import NaN
