+++ coloraide/color/interpolate.py
@@ -434,96 +434,3 @@ def color_lerp(color1, color2, space, out_space, progress, hue, premultiplied):
         outspace=out_space,
         premultiplied=premultiplied
     )
-
-
-class Interpolate:
-    """Interpolate between colors."""
-
-    def mask(self, channel, *, invert=False, in_place=False):
-        """Mask color channels."""
-
-        this = self if in_place else self.clone()
-        aliases = self._space.CHANNEL_ALIASES
-        masks = set(
-            [aliases.get(channel, channel)] if isinstance(channel, str) else [aliases.get(c, c) for c in channel]
-        )
-        for name in self._space.CHANNEL_NAMES:
-            if (not invert and name in masks) or (invert and name not in masks):
-                this.set(name, util.NaN)
-        return this
-
-    def steps(self, color, *, steps=2, max_steps=1000, max_delta_e=0, **interpolate_args):
-        """
-        Discrete steps.
-
-        This is built upon the interpolate function, and will return a list of
-        colors containing a minimum of colors equal to `steps` or steps as specified
-        derived from the `max_delta_e` parameter (whichever is greatest).
-
-        Number of colors can be capped with `max_steps`.
-
-        Default delta E method used is delta E 76.
-        """
-
-        return self.interpolate(color, **interpolate_args).steps(steps, max_steps, max_delta_e)
-
-    def mix(self, color, percent=util.DEF_MIX, *, in_place=False, **interpolate_args):
-        """
-        Mix colors using interpolation.
-
-        This uses the interpolate method to find the center point between the two colors.
-        The basic mixing logic is outlined in the CSS level 5 draft.
-        """
-
-        if not self._is_color(color) and not isinstance(color, (str, Piecewise, Mapping)):
-            raise TypeError("Unexpected type '{}'".format(type(color)))
-        color = self.interpolate(color, **interpolate_args)(percent)
-        return self.mutate(color) if in_place else color
-
-    def interpolate(
-        self, color, *, space="lab", out_space=None, stop=0, progress=None, hue=util.DEF_HUE_ADJ, premultiplied=False
-    ):
-        """
-        Return an interpolation function.
-
-        The function will return an interpolation function that accepts a value (which should
-        be in the range of [0..1] and will return a color based on that value.
-
-        While we use NaNs to mask off channels when doing the interpolation, we do not allow
-        arbitrary specification of NaNs by the user, they must specify channels via `adjust`
-        if they which to target specific channels for mixing. Null hues become NaNs before
-        mixing occurs.
-        """
-
-        space = space.lower()
-        out_space = self.space() if out_space is None else out_space.lower()
-
-        # A piecewise object was provided, so treat it as such,
-        # or we've changed the stop of the base color, so run it through piecewise.
-        if (
-            isinstance(color, Piecewise) or
-            (stop != 0 and (isinstance(color, str) or self._is_color(color)))
-        ):
-            color = [color]
-
-        if not isinstance(color, str) and isinstance(color, Sequence):
-            # We have a sequence, so use piecewise interpolation
-            return color_piecewise_lerp(
-                [Piecewise(self, stop=stop)] + list(color),
-                space,
-                out_space,
-                progress,
-                hue,
-                premultiplied
-            )
-        else:
-            # We have a sequence, so use piecewise interpolation
-            return color_lerp(
-                self,
-                color,
-                space,
-                out_space,
-                progress,
-                hue,
-                premultiplied
-            )
